some love for xaml namespaces, move files
