--- Src/MoneyFox.Windows/ApplicationInsights.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<ApplicationInsights xmlns="http://schemas.microsoft.com/ApplicationInsights/2013/Settings">
-  <!-- This key is for Application Insights resource 'MoneyManager' in resource group 'Default-ApplicationInsights-CentralUS' -->
-  <InstrumentationKey>ac915a37-36f5-436a-b85b-5a5617838bc8</InstrumentationKey>
-</ApplicationInsights>
