+++ Src/MoneyFox.Windows/Views/AppShell.xaml
@@ -1,11 +1,11 @@
 ﻿<Page x:Name="Root"
-      x:Class="MoneyFox.Windows.AppShell"
+      x:Class="MoneyFox.Windows.Views.AppShell"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:local="using:MoneyManager.Windows"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       xmlns:controls="using:MoneyFox.Windows.Controls"
+      xmlns:windows="using:MoneyFox.Windows"
       Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
       KeyDown="AppShell_KeyDown"
       TabNavigation="Cycle"
@@ -19,7 +19,7 @@
     -->
 
     <Page.Resources>
-        <DataTemplate x:Key="NavMenuItemTemplate" x:DataType="local:NavMenuItem">
+        <DataTemplate x:Key="NavMenuItemTemplate" x:DataType="windows:NavMenuItem">
             <Grid>
                 <Grid.ColumnDefinitions>
                     <ColumnDefinition Width="48" />
