+++ Src/MoneyFox.Windows/Views/BackupView.xaml
@@ -6,7 +6,7 @@
       xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
       xmlns:core="using:Microsoft.Xaml.Interactions.Core"
       xmlns:converters="using:Cimbalino.Toolkit.Converters"
-      xmlns:userControls1="using:MoneyFox.Windows.Views.UserControls"
+      xmlns:userControls="using:MoneyFox.Windows.Views.UserControls"
       mc:Ignorable="d"
       Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
@@ -26,11 +26,11 @@
             <RowDefinition Height="*" />
         </Grid.RowDefinitions>
 
-        <userControls1:PageHeader>
-            <userControls1:PageHeader.HeaderContent>
+        <userControls:PageHeader>
+            <userControls:PageHeader.HeaderContent>
                 <TextBlock Style="{ThemeResource PageTitleTextBlockStyle}" Text="Backup" x:Uid="BackupTitle" />
-            </userControls1:PageHeader.HeaderContent>
-        </userControls1:PageHeader>
+            </userControls:PageHeader.HeaderContent>
+        </userControls:PageHeader>
 
         <StackPanel Grid.Row="1"
                     Margin="19,9.5,19,0"
