+++ Src/MoneyFox.Windows/Views/Dialogs/SelectDateRangeDialog.xaml
@@ -3,7 +3,7 @@
                xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-               xmlns:converter1="using:MoneyFox.Windows.Converter"
+               xmlns:converter="using:MoneyFox.Windows.Converter"
                mc:Ignorable="d"
                Title="Select Daterange"
                x:Uid="SelectDateRangeDialog"
@@ -12,7 +12,7 @@
                PrimaryButtonCommand="{Binding DoneCommand}">
 
     <ContentDialog.Resources>
-        <converter1:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
+        <converter:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
     </ContentDialog.Resources>
