+++ Src/MoneyFox.Windows/Views/ModifyPaymentView.xaml
@@ -6,20 +6,19 @@
       xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
       xmlns:core="using:Microsoft.Xaml.Interactions.Core"
       xmlns:converters="using:Cimbalino.Toolkit.Converters"
-      xmlns:model="using:MoneyFox.Foundation.Model"
-      xmlns:converter1="using:MoneyFox.Windows.Converter"
-      xmlns:userControls1="using:MoneyFox.Windows.Views.UserControls"
-      xmlns:model1="using:MoneyFox.Core.Model"
+      xmlns:converter="using:MoneyFox.Windows.Converter"
+      xmlns:userControls="using:MoneyFox.Windows.Views.UserControls"
+      xmlns:model="using:MoneyFox.Core.Model"
       mc:Ignorable="d"
       NavigationCacheMode="Enabled"
       Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
     <Page.Resources>
-        <converter1:AmountFormatConverter x:Key="AmountConverter" />
-        <converter1:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
+        <converter:AmountFormatConverter x:Key="AmountConverter" />
+        <converter:DateTimeToDateTimeOffsetConverter x:Key="DateTimeToDateTimeOffsetConverter" />
         <converters:BooleanToVisibilityConverter x:Key="BooleanVisibilityConverter" />
         <DataTemplate x:Key="ComboBoxTemplate"
-                      x:DataType="model1:Account">
+                      x:DataType="model:Account">
             <StackPanel Margin="0,0,0,12"
                         Orientation="Horizontal">
                 <TextBlock Text="{x:Bind Name}" />
@@ -37,13 +36,13 @@
             <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
 
-        <userControls1:PageHeader>
-            <userControls1:PageHeader.HeaderContent>
+        <userControls:PageHeader>
+            <userControls:PageHeader.HeaderContent>
                 <TextBlock Style="{ThemeResource PageTitleTextBlockStyle}"
                            Text="Add Payment"
                            x:Uid="ModifyPaymentViewTitle" />
-            </userControls1:PageHeader.HeaderContent>
-        </userControls1:PageHeader>
+            </userControls:PageHeader.HeaderContent>
+        </userControls:PageHeader>
 
         <Grid Grid.Row="1"
               EntranceNavigationTransitionInfo.IsTargetElement="True">
