+++ Src/MoneyFox.Windows/Views/UserControls/CategoryListUserControl.xaml
@@ -6,9 +6,8 @@
              xmlns:designTime="using:MoneyManager.Core.ViewModels.DesignTime"
              xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
              xmlns:core="using:Microsoft.Xaml.Interactions.Core"
-             xmlns:model1="using:MoneyFox.Foundation.Model"
-             xmlns:controls1="using:MoneyFox.Windows.Controls"
-             xmlns:converter1="using:MoneyFox.Windows.Converter"
+             xmlns:controls="using:MoneyFox.Windows.Controls"
+             xmlns:converter="using:MoneyFox.Windows.Converter"
              xmlns:model="using:MoneyFox.Core.Model"
              d:DataContext="{d:DesignInstance designTime:DesignTimeCategoryListViewModel, IsDesignTimeCreatable=True}"
              mc:Ignorable="d"
@@ -16,7 +15,7 @@
              d:DesignWidth="400">
 
     <UserControl.Resources>
-        <converter1:ClickConverter x:Key="ClickConverter" />
+        <converter:ClickConverter x:Key="ClickConverter" />
 
         <DataTemplate x:Key="CategoryTemplate"
                       x:DataType="model:Category">
@@ -52,7 +51,7 @@
                  x:Uid="SearchHeader"
                  Text="{Binding SearchText, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
         <ScrollViewer Margin="0,7,0,0" Grid.Row="1">
-            <controls1:AlternatingRowListView OddRowBackground="{StaticResource ListViewBackgroundOddBrush}"
+            <controls:AlternatingRowListView OddRowBackground="{StaticResource ListViewBackgroundOddBrush}"
                                               EvenRowBackground="{StaticResource ListViewBackgroundEvenBrush}"
                                               ItemsSource="{Binding Categories}"
                                               ItemTemplate="{StaticResource CategoryTemplate }"
@@ -71,7 +70,7 @@
                                 Value="Stretch" />
                     </Style>
                 </ListView.ItemContainerStyle>
-            </controls1:AlternatingRowListView>
+            </controls:AlternatingRowListView>
         </ScrollViewer>
     </Grid>
 </UserControl>
